Move session struct to new session diractory
